+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Driver swapping via factories
 - Quoting mode
-
+- Laravel integration
 
 ## [0.2.0] - 2024-05-27
 ### Added